Use AKImageViewerViewController as an example to configure my own working ImageViewerViewController.
When a user taps on a picture, the user is taken to another view, where
they can zoom into the picture and save the picture. Most of the
functionalities are set. Right now, I am in the process of configuring
the functionality for dismissing the picture enlarged view when the
user swipes upward or downward.